Add a new feature for hiding table border. Toggle menu item within the option menu. Hiding is also a undoable/redoable property change. We use hidden tables to organized column data.
Within the feature we can also resurrect tables that start with the hidden class.

Since we start table construction from cell to table within the blots, you'll see matcher will contain whether the table element brought in contains the hidden class. in the blot you'll see we utilize it when we start a table up in the blot.

Toolbar recgonizes a table border has been shown or hidden and shows the user the appropriate toggle.

Everything seemed to test and work very well. Let me know if this also looks good on your end and/or you have suggestions.